requirements.txt
fastapi==0.121.2
uvicorn[standard]==0.38.0
sqlalchemy==2.0.44
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
psycopg2-binary==2.9.11
pydantic[email]==2.12.4
bcrypt==4.1.2
